make http-json:integration-tests into test suites
