+++ ledger-service/http-json/BUILD.bazel
@@ -7,6 +7,7 @@ load(
     "da_scala_binary",
     "da_scala_library",
     "da_scala_test",
+    "da_scala_test_suite",
     "lf_scalacopts",
     "silencer_plugin",
 )
@@ -341,7 +342,81 @@ alias(
 )
 
 [
-    da_scala_test(
+    da_scala_library(
+        name = "integration-tests-lib-postgres-{}".format(edition),
+        srcs = glob([
+            "src/itpostgres/scala/**/*.scala",
+        ]),
+        plugins = [
+            "@maven//:org_typelevel_kind_projector_{}".format(scala_version_suffix),
+            silencer_plugin,
+        ],
+        scala_deps = [
+            "@maven//:com_chuusai_shapeless",
+            "@maven//:com_typesafe_akka_akka_http_core",
+            "@maven//:com_typesafe_scala_logging_scala_logging",
+            "@maven//:io_spray_spray_json",
+            # "@maven//:org_scalacheck_scalacheck",
+            "@maven//:org_scalactic_scalactic",
+            "@maven//:org_scalatest_scalatest_core",
+            # "@maven//:org_scalatest_scalatest_flatspec",
+            # "@maven//:org_scalatest_scalatest_freespec",
+            "@maven//:org_scalatest_scalatest_matchers_core",
+            "@maven//:org_scalatest_scalatest_shouldmatchers",
+            "@maven//:org_scalaz_scalaz_core",
+            # "@maven//:org_tpolecat_doobie_core",
+            # "@maven//:org_tpolecat_doobie_free",
+            # "@maven//:org_tpolecat_doobie_postgres",
+            # "@maven//:org_typelevel_cats_core",
+            # "@maven//:org_typelevel_cats_effect",
+            # "@maven//:org_typelevel_cats_free",
+            # "@maven//:org_typelevel_cats_kernel",
+            # "@maven//:org_scala_lang_modules_scala_collection_compat",
+        ],
+        scalacopts = hj_scalacopts,
+        deps = [
+            ":http-json-{}".format(edition),
+            ":integration-tests-lib-{}".format(edition),
+            # "//bazel_tools/runfiles:scala_runfiles",
+            "//daml-lf/data",
+            "//daml-lf/interface",
+            "//daml-lf/transaction",
+            "//daml-lf/transaction-test-lib",
+            "//language-support/scala/bindings-akka",
+            "//ledger-api/rs-grpc-bridge",
+            "//ledger-api/testing-utils",
+            # "//ledger-service/db-backend",
+            "//ledger-service/fetch-contracts",
+            "//ledger/ledger-api-auth",
+            "//ledger/ledger-resources",
+            "//ledger/metrics",
+            "//ledger/sandbox:sandbox-scala-tests-lib",
+            "//ledger/sandbox-common",
+            "//ledger/sandbox-common:sandbox-common-scala-tests-lib",
+            "//ledger-service/http-json-cli:{}".format(edition),
+            "//ledger-service/http-json-testing:{}".format(edition),
+            "//ledger-service/jwt",
+            "//ledger-service/utils",
+            # "//libs-scala/contextualized-logging",
+            "//libs-scala/db-utils",
+            "//libs-scala/ports",
+            "//libs-scala/postgresql-testing",
+            # "//libs-scala/resources",
+            # "//libs-scala/scala-utils",
+            # "//runtime-components/non-repudiation",
+            # "//runtime-components/non-repudiation-postgresql",
+            "@maven//:io_dropwizard_metrics_metrics_core",
+            "@maven//:org_scalatest_scalatest_compatible",
+        ],
+    )
+    for edition in [
+        "ce",
+        "ee",
+    ]
+]
+
+[
+    da_scala_test_suite(
         name = "integration-tests-{}".format(edition),
         size = "large",
         srcs = glob([
@@ -391,6 +466,7 @@ alias(
         deps = [
             ":http-json-{}".format(edition),
             ":integration-tests-lib-{}".format(edition),
+            ":integration-tests-lib-postgres-{}".format(edition),
             "//bazel_tools/runfiles:scala_runfiles",
             "//daml-lf/data",
             "//daml-lf/interface",
